remove NodataReader
When an `open` fails and we'd return nodata instead, we don't cache that anymore. So if the asset doesn't exist, we'll try to open it for every chunk. Unclear how much of a performance impact this will have. It's probably not ideal, because combining it with retries could be nice: retry errors first, then give up and use nodata if they persist?
